Now watches for process instead of window
Deciding whether or not a tool is already up by a window existing fails
if the tool is started, but hasn’t created a window yet. This is not an
issue if watching for the _process_ to pop up.

Ideally this would also close processes instead of windows when the
watched game goes down, but killing a process from AHK sadly is
ungraceful, unlike closing a window.